+++ packages/test-utils/mocha-hooks/docker.js
@@ -4,6 +4,9 @@ const checkForDockerHostname = require('../src/check-for-docker-hostname')
 exports.mochaHooks = {
   async beforeAll() {
     await checkForDockerHostname()
-    await waitForDockerBrowsers()
+    await waitForDockerBrowsers({
+      remoteUrl: process.env.CVG_TESTS_REMOTE,
+      retries: Number(process.env.CVG_TESTS_DOCKER_RETRY_COUNT) || 70,
+    })
   },
 }