+++ packages/test-utils/src/wait-for-docker-browsers.js
@@ -2,13 +2,12 @@ const {promisify} = require('util')
 const fetch = require('node-fetch')
 const delay = promisify(setTimeout)
 
-async function waitForDockerBrowsers({
-  remoteUrl = process.env.CVG_TESTS_REMOTE,
-  retries = Number(process.env.CVG_TESTS_DOCKER_RETRY_COUNT) || 70,
-} = {}) {
-  if (retries === 0) {
-    throw new Error('browsers docker containers failed to start before running tests')
-  }
+async function waitForDockerBrowsers({remoteUrl, retries} = {}) {
+  if (!remoteUrl)
+    throw new Error(
+      'URL not provided for container health check (e.g., typically through the CVG_TESTS_REMOTE env var)',
+    )
+  if (retries === 0) throw new Error('browsers docker containers failed to start before running tests')
   try {
     await fetch(remoteUrl)
   } catch (_ex) {