+++ packages/test-utils/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- add error handling when a URL is not provided when checking if a docker container is ready
 
 ## 1.1.0 - 2022/3/2